One of the twelve was in a file the first sweep never saw
CodeQL alert 44 was still open after the format-string commit, and it was right:
`kernel/arch/riscv64/mmu.c` printed `(uint64_t)g_early_used` through `%u`.

The reason it survived is worth recording. The two searches that produced the
list found different sets. The compiler sweep used the x86_64 file set, which
excludes `arch/riscv64` by construction, so it could not see this call at all;
the ten it did find were a different ten from CodeQL's. Between them the union
is twelve call sites, and only eleven were fixed.

The fix is the one `user.c` needed rather than a widening: `g_early_used` is a
`uint32_t`, so `%u` was already the right conversion and the cast was the thing
that broke the agreement between specifier and argument.

Not verified locally: this machine's working copy sits in Dropbox and its
FileProvider cannot read several headers, `spinlock.h` among them, so nothing
that includes `smp.h` compiles here at all -- `fsyntax-only` on this file fails
with "cannot open file" before it reaches the call. `compile-check` on the
runner covers all three architectures on a clean checkout and is what settles
it.

+++ kernel/arch/riscv64/mmu.c
@@ -131,7 +131,7 @@ static uint64_t *early_table(void) {
        with nothing connecting the two. Under UEFI that was a store to the
        interrupt controller, twenty log lines after the mapping that failed. */
     klog("vmm: early page-table pool exhausted after %u tables; some of the "
-         "identity map was not created\n", (uint64_t)g_early_used);
+         "identity map was not created\n", g_early_used);
     return 0;
   }
   uint64_t *table = g_early[g_early_used++];
